feat(details): implement purchase history display and add ViewModel for data handling

+++ app/src/main/java/me/kire/re/homestuffapp/data/remote/dto/Content.kt
@@ -12,7 +12,7 @@ data class Content(
 ) {
     fun toNourishment(): Product {
         return Product(
-            productId = this.nourishmentId,
+            productId = this.nourishmentId.toLong(),
             name = this.name,
             imageUrl = this.imageUrl,
             description = this.description,
